Структурирование проекта, для дальнейшего запуска и интегрирование Telegram бота и WebApp в Telegram Mini Apps.
