+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__/
 .env
 .DS_Store
 .pytest_cache/
+!/log/log.log
