+++ README.md
@@ -1,4 +1,4 @@
-# **Проект "Workout Tracker"**
+# **Проект "StrongTrack_Bot"**
 
 "Workout Tracker" — это веб-приложение для учета тренировок. Пользователь может выбрать день недели, указать количество
 подходов, вес и количество повторений, после чего приложение рассчитает общий поднятый вес за тренировку.
