+++ TODO.md
@@ -1,3 +1,13 @@
 ~~1. Создать страниц для всех функций WebApp.~~
-1. Добавить в боте, вызов WebApp.
-2. Источник: https://powermens.ru/?modul=catfiles&p=more&idfile=141
+
+1. Добавить в боте вызов WebApp.
+2. Источник: https://powermens.ru/?modul=catfiles&p=more&idfile=141.
+3. Переписать клавиатуры на InlineKeyboardMarkup c ReplyKeyboardMarkup, так как из под ReplyKeyboardMarkup.
+4. Проверка пользователей должна быть по user_id = message.from_user.id, так как не у всех есть username.
+5. При вводе веса должна быть проверка, чтобы значения были цифрами, а не словами.
+6. При вводе роста должно быть ограничение до разумных пределов (не допускать, например, 5 метров и 40 см).
+7. При вводе опыта тренировок нужно добавить диапазон, например: 1–2 года, 2–5 лет и так далее.
+8. При запуске веб-апп через https://ngrok.com/, продумать более простой запуск бота, чтобы поочередно не запускать проекты, возможно, в будущем будем использовать докер.
+9. Для более хорошей гибкости работы проекта и управления текстовыми данными, по возможности выводить данные в json-файлы.
+10. Продумать хорошую структуру проекта для лучшей масштабируемости.
+11. Доработать документацию README.md с учетом изменениями проекта.
