+++ WebApp/templates/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Дельтовидные мышцы</title>
+    <title>Главная страница</title>
     <style>
         /* Сброс стандартных стилей */
         * {
@@ -137,11 +137,7 @@
             border-radius: 8px;
             box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
             padding: 20px;
-            text-align: left; /* Выравнивание текста по левому краю */
-        }
-
-        p, h1, h2, h3, b, i {
-            text-align: left;
+            text-align: center;
         }
 
         h1 {
@@ -276,62 +272,35 @@
         <div class="menu-button"></div>
     </label>
     <ul class="menu">
-        <li><a href="../../index.html">Главная</a></li>
-        <li><a href="../../about_us.html">О нас</a></li>
-        <li><a href="../../contacts.html">Контакты</a></li>
-        <li><a href="../../settings.html">Настройки</a></li>
+        <li><a href="index.html">Главная</a></li>
+        <li><a href="about_us.html">О нас</a></li>
+        <li><a href="contacts.html">Контакты</a></li>
+        <li><a href="settings.html">Настройки</a></li>
     </ul>
 </section>
 
 <!-- Контейнер с контентом -->
 <div class="container">
 
-    <h1>Дельтовидные мышцы</h1>
-
-    <div style="text-align: center;">
-        <img src="dumbbell_fly_bentover.jpg" alt="Дельтовидные мышцы" width="400">
+    <!-- Информация о пользователе -->
+    <div class="profile-info">
+        <!--        Размещаем изображение с помощью FastAPI-->
+        <img src="{{ url_for('static', path='images/avatar.jpg') }}" alt="profile-picture" class="profile-picture">
+        <!--        <img src="../static/images/avatar.jpg" alt="Profile Picture" class="profile-picture">-->
+        <h3>Шварц</h3>
+        <p class="phone-number">8876116536</p>
     </div>
 
-    <h2>Разведение гантелей в наклоне</h2><br>
-
-    <i>Задние дельты, трапеции и мышцы-вращатели плеча /Формирующее упражнение/ Форма и рельеф задних плеч</i><br><br>
-
-    <b>Разведения в наклоне</b> — изолируют нагрузку на заднюю головку дельт и применяются для того, чтобы придать этому пучку мышц отчетливую бугристую форму, добавить «полосатости» и выделить его на фоне мышц спины.<br>
-
-    Включив разведения в наклоне в программу силовой подготовки, вы улучшите спортивные достижения во всех видах спорта, в которых вы часто подтягиваете руки к телу (плавание, гимнастика) или отводите их назад (стрельба из лука, гребля), совершаете удары руками сверху вниз (волейбол, теннис) или удерживаете соперника возле себя (борьба).<br><br>
-
-    Запомните: чем сильнее задние дельты, тем крепче плечевой сустав.<br>
-
-    <h3>Техника: Разведение гантелей в наклоне</h3><br>
-    1. Возьмите легкие гантели в обе руки нейтральным хватом (ладони смотрят друг на друга).<br>
-    2. Наклонитесь вперед так, чтобы торс был параллелен полу. Спина ровная, чуть прогнута в пояснице.<br>
-    3. Если у вас тугие мышцы задней части бедра, можно слегка согнуть ноги в коленях. Главное — прочно зафиксировать нормальный, S-образный изгиб позвоночника. Но если и это не помогает и вам все-таки тяжело держать торс в таком положении, воспользуйтесь горизонтальной скамьей. Расположитесь на ней так, чтобы почти вся грудь «вылазила» за ее край (скамья не должна мешать разводить гантели), а ноги прочно упирались в пол.<br>
-    4. В исходном положении гантели свисают на выпрямленных руках под грудью. Руки должны быть прочно зафиксированы в локтях до самого конца сета. Их можно чуть-чуть согнуть, перед тем как приступить к выполнению упражнения, но сгибать или разгибать их во время сета не допускается. Все движение происходит только в плечевых суставах, все остальные суставы — неподвижны!<br>
-    5. Сделайте вдох и задержите дыхание. Изо всех сил напрягите задние дельты и трапеции и плавно разведите гантели в стороны, стараясь поднять их как можно выше.<br>
-    6. Обратите особое внимание: руки должны двигаться исключительно в вертикальной плоскости, проходящей через плечи. Не отводите гантели ни назад, ни вперед.<br>
-    7. В верхней точке локти должны быть выше уровня спины.<br>
-    8. Выдохните и плавно опустите гантели в исходное положение. Сделайте короткую паузу и выполните следующее повторение.<br>
-
-    <h3>Советы: Разведение гантелей в наклоне</h3><br>
-    1. Всегда держите спину ровной, чуть прогнутой в пояснице. Даже легкое скругление спины чревато травмой.<br>
-    2. Задержка дыхания во время разведения рук не только стабилизирует положение торса, но и делает вас сильнее.<br>
-    3. Чтобы нагрузить задние дельты по максимуму, опускайте плечи в исходном положении и поднимайте гантели как можно выше уровня спины, но при этом не сводите лопатки вместе.<br>
-    4. Чтобы прицельно отработать средние трапеции и ромбовидные мышцы и добиться их максимального сокращения, обязательно сводите лопатки в верхней точке.<br>
-    5. Не используйте чрезмерно тяжелые гантели. Они могут не дать вам отработать упражнение технически правильно и стать причиной травмы.<br>
-    6. Держите руки выпрямленными или чуть согнутыми в локтях. Имейте в виду: чем больше руки согнуты в локтях, тем короче амплитуда движения и тем меньше нагрузка на задние дельты.<br>
-    7. Вариация упражнения — перекрестные разведения в кроссовере. Поставьте скамью посередине кроссовера так, чтобы ваши плечи оказались в одной вертикальной плоскости с нижними блоками. Прикрепите D-образные рукоятки к тросам, проходящим через эти блоки, и, скрещивая тросы, выполняйте разведения. Однако кроссовер, по сравнению с гантелями, имеет существенный недостаток — вы не можете выполнять разведения по полной амплитуде, так как развести рукоятки выше уровня плеч вам мешают перекрещивающиеся тросы.<br>
-    8. Еще одна вариация — разведения одной рукой в блочном тренажере стоя на коленях. В исходном положении вы становитесь на четвереньки, упираетесь одной рукой в пол и слегка сгибаете ее так, чтобы плечо рабочей руки оказалось приподнятым (иначе во время возвращения в исходную позицию рукоятка будет цепляться за пол).<br><br>
-
-    <h3>Применение: Разведение гантелей в наклоне</h3><br>
-
-    <b>Кому:</b> Всем, от новичка до мастера.<br><br>
-
-    <b>Когда:</b> В конце тренировки плеч. Перед разведениями гантелей в наклоне сделайте жим гантелей/штанги сидя/стоя, разведения гантелей стоя и подъемы перед собой.<br><br>
-
-    <b>Сколько:</b> 3-4 сета по 8-12 повторений.<br><br>
+    <p>С этим Telegram-ботом вы добьетесь небывалых высот и станете настоящей звездой!</p>
 
     <!-- Кнопки -->
-    <button class="button2"><a href="../../index.html">В главное меню</a></button>
+    <button class="button2"><a href="view_results.html">Посмотреть результаты</a></button>
+    <button class="button2"><a href="view_workouts.html">Посмотреть тренировки</a></button>
+    <button class="button2"><a href="сreate_workouts.html">Сформировать тренировки</a></button>
+    <button class="button2"><a href="my_details.html">Мои данные</a></button>
+    <button class="button2"><a href="basic_exercises.html">Базовые упражнения</a></button>
+    <button class="button2"><a href="nutrition.html">Питание</a></button>
+
 
     <div class="button-container">
         <button class="button">
