+++ WebApp/templates/types_of_exercises.html
