+++ data/text.py
@@ -6,7 +6,7 @@ def text_hello_welcome() -> str:
     """
     text = (
         "\n\n💪 Хочу познакомить тебя с нашим проектом!\n\n"
-        "'🏋️ Workout Tracker' — это веб-приложение для учета тренировок. "
+        "'🏋️ StrongTrack_Bot' — это веб-приложение для учета тренировок. "
         "Пользователь может выбрать день недели, указать количество подходов, вес и количество повторений, "
         "после чего приложение рассчитает общий поднятый вес за тренировку."
     )
@@ -36,7 +36,7 @@ def text_authorized_user_greeting() -> str:
     :return: Строка с приветствием авторизованного пользователя.
     """
     text = (
-        "\n\n'🏋️ Workout Tracker' — это веб-приложение для учета тренировок. "
+        "\n\n'🏋️ StrongTrack_Bot' — это веб-приложение для учета тренировок. "
         "Пользователь может выбрать день недели, указать количество подходов, вес и количество повторений, "
         "после чего приложение рассчитает общий поднятый вес за тренировку."
     )
